chore: move traits into modules
