+++ crates/wallet/storage_sqlite/tests/substates.rs
@@ -5,7 +5,7 @@ use std::{collections::HashSet, str::FromStr};
 
 use tari_engine_types::substate::SubstateId;
 use tari_ootle_common_types::{optional::Optional, VersionedSubstateIdRef};
-use tari_ootle_wallet_sdk::storage::{CommitableStore, WalletStoreReader, WalletStoreWriter, WriteableWalletStore};
+use tari_ootle_wallet_sdk::storage::{CommittableStore, WalletStoreReader, WalletStoreWriter, WriteableWalletStore};
 use tari_ootle_wallet_storage_sqlite::SqliteWalletStore;
 
 #[test]
